Refactored the api routes to prepare for send to functionality
