+++ .gitignore
@@ -493,3 +493,6 @@ wwwroot/
 config/*.db
 config/*.db-shm
 config/*.db-wal
+KavitaEmail/config/*.db
+KavitaEmail/config/*.db-shm
+KavitaEmail/config/*.db-wal
