+++ KavitaEmail/Controllers/SendToController.cs
@@ -1,0 +1,9 @@
+﻿namespace Skeleton.Controllers;
+
+/// <summary>
+/// Responsible for all Send to Device Email Services
+/// </summary>
+public class SendToController : BaseApiController
+{
+    
+}
